backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "fastapi[standard]>=0.131.0",
    "google-genai>=1.64.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
